GH-1491: Move tests of old turtle parser to ttl_test package
