+++ jena-core/src/test/java/org/apache/jena/ttl_test/TupleItem.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.jena.ttl;
+package org.apache.jena.ttl_test;
 
 /**
  * The unit found in a line of a tuple.
