+++ jena-core/src/test/java/org/apache/jena/ttl_test/TupleSet.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.jena.ttl ;
+package org.apache.jena.ttl_test ;
 
 import java.io.* ;
 import java.util.* ;
