+++ jena-core/src/test/java/org/apache/jena/ttl_test/turtle/TestBadSyntax.java
@@ -16,24 +16,31 @@
  * limitations under the License.
  */
 
-package org.apache.jena.ttl.turtle;
+package org.apache.jena.ttl_test.turtle;
 
-import junit.framework.* ;
+import junit.framework.TestCase;
+import org.apache.jena.rdf.model.* ;
+import org.apache.jena.ttl.turtle.TurtleParseException;
+import org.apache.jena.ttl.turtle.TurtleReader;
 
-public class TurtleTestSuite extends TestSuite
+
+public class TestBadSyntax extends TestCase
 {
-    static public TestSuite suite() {
-        return new TurtleTestSuite() ;
+    String uri ;
+    public TestBadSyntax(String name, String uri) { super(name) ; this.uri = uri ; }
+    
+    @Override
+    public void runTest()
+    {
+        Model model = ModelFactory.createDefaultModel() ;
+        RDFReaderI t = new TurtleReader() ;
+        try {
+            t.read(model, uri) ;
+            fail("Bad syntax test succeed in parsing the file") ;
+        } catch (TurtleParseException ex)
+        {
+            return ;    
+        }
+
     }
-	
-	private TurtleTestSuite()
-	{
-		super("Turtle") ;
-        addTest(TestTurtleReader.suite()) ;
-        addTest(TurtleInternalTests.suite()) ;
-        addTest(TurtleTestFactory.make("testing/Turtle/manifest.ttl")) ;
-//		addTest(new N3ExternalTests()) ;
-//		addTest(new N3JenaReaderTests()) ;
-//		addTest(new N3JenaWriterTests()) ;
-	}
 }
