+++ jena-core/src/test/java/org/apache/jena/ttl_test/turtle/TurtleTestFactory.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.jena.ttl.turtle;
+package org.apache.jena.ttl_test.turtle;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
