+++ jena-core/src/test/java/org/apache/jena/ttl_test/turtle/TurtleTestVocab.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.jena.ttl.turtle ;
+package org.apache.jena.ttl_test.turtle ;
 
 import org.apache.jena.rdf.model.* ;
